requirements.txt
hidapi
docutils